@@ -1,21 +1,22 @@
 module github.com/digisan/cli-prompt
 
-go 1.17
+go 1.19
 
 require (
-	github.com/digisan/go-generics v0.0.8
-	github.com/digisan/gotk v0.0.46
-	github.com/digisan/json-tool v0.0.33
-	github.com/tidwall/sjson v1.2.4
+	github.com/digisan/go-generics v0.2.3
+	github.com/digisan/gotk v0.1.23
+	github.com/digisan/json-tool v0.1.6
+	github.com/tidwall/sjson v1.2.5
 )
 
 require (
 	github.com/clbanning/mxj v1.8.4 // indirect
-	github.com/digisan/logkit v0.0.4 // indirect
-	github.com/gookit/color v1.5.0 // indirect
-	github.com/tidwall/gjson v1.12.1 // indirect
+	github.com/digisan/logkit v0.1.2 // indirect
+	github.com/gookit/color v1.5.2 // indirect
+	github.com/tidwall/gjson v1.14.3 // indirect
 	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.0 // indirect
-	github.com/xo/terminfo v0.0.0-20210125001918-ca9a967f8778 // indirect
-	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e // indirect
+	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
+	golang.org/x/net v0.0.0-20220921203646-d300de134e69 // indirect
+	golang.org/x/sys v0.0.0-20220919091848-fb04ddd9f9c8 // indirect
 )
